planner: binding from cluster statements summary (#39737)
ref pingcap/tidb#39199
