+++ executor/tiflashtest/BUILD.bazel
@@ -15,7 +15,6 @@ go_test(
         "//domain",
         "//executor",
         "//meta/autoid",
-        "//parser/auth",
         "//parser/terror",
         "//store/mockstore",
         "//store/mockstore/unistore",
