+++ infoschema/BUILD.bazel
@@ -93,8 +93,10 @@ go_test(
         "//store/helper",
         "//store/mockstore",
         "//store/mockstore/mockstorage",
+        "//store/mockstore/unistore",
         "//table",
         "//testkit",
+        "//testkit/external",
         "//testkit/testsetup",
         "//testkit/testutil",
         "//types",
@@ -104,14 +106,17 @@ go_test(
         "//util/pdapi",
         "//util/resourcegrouptag",
         "//util/set",
+        "//util/stmtsummary",
         "@com_github_gorilla_mux//:mux",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_pingcap_fn//:fn",
         "@com_github_pingcap_kvproto//pkg/deadlock",
+        "@com_github_pingcap_kvproto//pkg/metapb",
         "@com_github_pingcap_tipb//go-tipb",
         "@com_github_prometheus_prometheus//promql",
         "@com_github_stretchr_testify//require",
+        "@com_github_tikv_client_go_v2//testutils",
         "@org_golang_google_grpc//:grpc",
         "@org_uber_go_goleak//:goleak",
     ],
