+++ util/stmtsummary/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//parser/mysql",
         "//sessionctx/stmtctx",
         "//types",
+        "//util/chunk",
         "//util/execdetails",
         "//util/hack",
         "//util/kvcache",
